+++ test/parameterized_tendencies/microphysics/allocations.jl
@@ -86,11 +86,10 @@ end
 
 function _allocs_cloud_fraction_simple()
     FT = Float64
-    CA._compute_cloud_fraction(FT(1e-3), FT(3.16e-4), FT(0), FT(5e-5), FT(1))
+    CA._compute_cloud_fraction(FT(1e-3), FT(3.16e-4), FT(5e-5), FT(1))
     return @allocated CA._compute_cloud_fraction(
         FT(1e-3),
         FT(3.16e-4),
-        FT(0),
         FT(5e-5),
         FT(1),
     )